Generalize the step input spec
Make the spec general purpose, and more suitable for use by Bolt
novices. This is the first step of a series of refactor iterations and
necessarily loses some functionality the previous iteration had. The
goal is to ensure the UX of the step is correct, and then iteratively
add back in any dropped functionality needed for real-world use cases.