Make the transport connection a oneOf specifier
This currently still only has one "oneOf" listed, but the intention is
to permit specifying WinRM transport details as an alternative to
specifying ssh private key details. Or, specifying a password. Point is,
we plan on permitting the specification of one of a small number of Bolt
supported transport types in this section.