Make .transport.connection optional
Previously, Bolt would always be configured to use an ssh key file. If a
connection is not specified, the key data does not exist, and Bolt will
error trying to parse the empty file.

This commit conditionally configures Bolt to use the file only if a data
value for the key is given.